requirements.txt
pylint
nose
read-only-property

